Stream now parses colour properly

+++ alpha-conferences/UIColor+Alpha.h
@@ -11,7 +11,7 @@
 @interface UIColor (Alpha)
 
 // utility method to create a UIColor object from a hexadecimal colour code
-+ (UIColor *)colorWithHex:(int)colorHex;
++ (UIColor *)colorWithHex:(NSUInteger)colorHex;
 
 // tint colour for all UINavigationBar objects
 + (UIColor *)navigationBarTintColour;
